feat: Optimize gas in router, update README with report, and add NatSpec

+++ test/ZapETH.t.sol
@@ -7,7 +7,7 @@ import "../src/Univ2ZapRouter.sol";
 contract ZapETH is Test {
     address constant ROUTER = 0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D;
     address constant WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
-    address constant USDC   = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
+    address constant USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
 
     Univ2ZapRouter zap;
     address user = vm.addr(11);
@@ -22,22 +22,28 @@ contract ZapETH is Test {
 
     function test_WETH_ZapInAndOut() public {
         vm.startPrank(user);
-        IERC20(WETH).approve(address(zap), type(uint).max);
+        IERC20(WETH).approve(address(zap), type(uint256).max);
 
-        uint lp = zap.zapInSingleToken(
+        uint256 lp = zap.zapInSingleToken(
             WETH,
-            USDC, WETH,
-            1 ether, 50, 1, block.timestamp + 1 hours,
+            USDC,
+            WETH,
+            1 ether,
+            50,
+            1,
+            block.timestamp + 1 hours,
             false // feeOnTransfer
         );
         assertGt(lp, 0);
 
         IERC20(zap.getPair(USDC, WETH)).approve(address(zap), lp);
-        uint wethOut = zap.zapOutSingleToken(
-            WETH, USDC, WETH,
+        uint256 wethOut = zap.zapOutSingleToken(
+            WETH,
+            USDC,
+            WETH,
             lp,
             50, // maxSlippageBps (0.5%)
-            1,  // outMin
+            1, // outMin
             block.timestamp + 1 hours,
             false // feeOnTransfer
         );
