+++ test/ZapFuzzUSDC.t.sol
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
+
 import "forge-std/Test.sol";
 import "../src/Univ2ZapRouter.sol";
 
 contract ZapFuzzUSDC is Test {
     address constant ROUTER = 0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D;
     address constant WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
-    address constant USDC   = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
+    address constant USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
 
     Univ2ZapRouter zap;
     address user = vm.addr(22);
@@ -17,20 +18,22 @@ contract ZapFuzzUSDC is Test {
     }
 
     function test_fuzz_USDC(uint96 amt) public {
-        amt = uint96(bound(amt, 100e6, 10_000e6));   // 100→10 000 USDC
+        amt = uint96(bound(amt, 100e6, 10_000e6)); // 100→10 000 USDC
         deal(USDC, user, amt);
 
         vm.startPrank(user);
-        IERC20(USDC).approve(address(zap), type(uint).max);
+        IERC20(USDC).approve(address(zap), type(uint256).max);
 
-        uint lp = zap.zapInSingleToken(USDC, USDC, WETH, amt, 50, 1, block.timestamp+1 hours, false);
+        uint256 lp = zap.zapInSingleToken(USDC, USDC, WETH, amt, 50, 1, block.timestamp + 1 hours, false);
         IERC20(zap.getPair(USDC, WETH)).approve(address(zap), lp);
-        uint out = zap.zapOutSingleToken(
-            USDC, USDC, WETH,
+        uint256 out = zap.zapOutSingleToken(
+            USDC,
+            USDC,
+            WETH,
             lp,
             50, // maxSlippageBps (0.5%)
-            1,  // outMin
-            block.timestamp+1 hours,
+            1, // outMin
+            block.timestamp + 1 hours,
             false // feeOnTransfer
         );
         vm.stopPrank();
