+++ test/ZapRoundTrip.t.sol
@@ -9,7 +9,7 @@ contract ZapRoundTrip is Test {
     /* ─── Mainnet constants (UniV2, USDC-WETH 0.3% pool) ───────── */
     address constant ROUTER = 0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D;
     address constant WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
-    address constant USDC   = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
+    address constant USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
 
     Univ2ZapRouter zap;
     address user = vm.addr(1);
@@ -27,7 +27,7 @@ contract ZapRoundTrip is Test {
     function testRoundTripUSDC() public {
         vm.startPrank(user);
 
-        IERC20(USDC).approve(address(zap), type(uint).max);
+        IERC20(USDC).approve(address(zap), type(uint256).max);
 
         /* ---- zap in USDC → LP ---- */
         // event ZapIn(
@@ -43,12 +43,13 @@ contract ZapRoundTrip is Test {
         vm.expectEmit(true, true, true, false);
         emit Univ2ZapRouter.ZapIn(user, USDC, USDC, WETH, 1_000e6, 0); // placeholder for liquidityMinted
 
-        uint lp = zap.zapInSingleToken(
-            USDC,              // tokenIn
-            USDC, WETH,        // pool tokens
-            1_000e6,           // amountIn
+        uint256 lp = zap.zapInSingleToken(
+            USDC, // tokenIn
+            USDC,
+            WETH, // pool tokens
+            1_000e6, // amountIn
             50,
-            1,                 // lpMin
+            1, // lpMin
             block.timestamp + 1 hours,
             false // feeOnTransfer
         );
@@ -70,11 +71,13 @@ contract ZapRoundTrip is Test {
         vm.expectEmit(true, true, true, false);
         emit Univ2ZapRouter.ZapOut(user, USDC, USDC, WETH, lp, 0); // placeholder for amountOut
 
-        uint usdcOut = zap.zapOutSingleToken(
-            USDC, USDC, WETH,
+        uint256 usdcOut = zap.zapOutSingleToken(
+            USDC,
+            USDC,
+            WETH,
             lp,
             50, // maxSlippageBps (0.5%)
-            1,  // outMin
+            1, // outMin
             block.timestamp + 1 hours,
             false // feeOnTransfer
         );
@@ -88,15 +91,16 @@ contract ZapRoundTrip is Test {
     function testRoundTripUSDC_FeeOnTransfer() public {
         vm.startPrank(user);
 
-        IERC20(USDC).approve(address(zap), type(uint).max);
+        IERC20(USDC).approve(address(zap), type(uint256).max);
 
         // Expect ZapIn event
         vm.expectEmit(true, true, true, false);
         emit Univ2ZapRouter.ZapIn(user, USDC, USDC, WETH, 1_000e6, 0);
 
-        uint lp = zap.zapInSingleToken(
+        uint256 lp = zap.zapInSingleToken(
             USDC,
-            USDC, WETH,
+            USDC,
+            WETH,
             1_000e6,
             50,
             1,
@@ -111,8 +115,10 @@ contract ZapRoundTrip is Test {
         vm.expectEmit(true, true, true, false);
         emit Univ2ZapRouter.ZapOut(user, USDC, USDC, WETH, lp, 0);
 
-        uint usdcOut = zap.zapOutSingleToken(
-            USDC, USDC, WETH,
+        uint256 usdcOut = zap.zapOutSingleToken(
+            USDC,
+            USDC,
+            WETH,
             lp,
             50,
             1,
